+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - replace lodash with lodash-es
 - update react 16.3.1
 - remove flow dependency
+- react router 4 
+	- update using [README](./src/modules/react-redux-hash-router4/doc/README.md)
 
 ## [0.5.1 ]